requirements.txt
coveralls
flake8
image
imutils
libusb
libusb1
mock
numpy
opencv-python
Pillow
pygame
pylint
pyopengl
pyopengl_accelerate
pyqt5
pyqt5-sip
pyserial
pytest
pytest-cov
pytest-shutil
pytorch
pyudev
pyusb
pyzbar
requests_mock
torch
torchvision
tox
